Support single string provide include as same as bare string
Found this missing case when writing docs. Specifiying `include:` as a single
string should be equivalent to specifying a bare string, which is already
treated as a one-element array.

Renaming only happens when as: is present, and we should support the
`include:` single string like a 1-element array, like the bare string case.

Signed-off-by: Robert Detjens <github@detjens.dev>